backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
neo4j==5.14.1
faker==20.1.0
python-dotenv==1.0.0
email-validator==2.1.0
openai==1.37.0
google-generativeai==0.7.2
pandas==2.2.2
pyarrow==15.0.2